Implement invalid field handling for user register

+++ src/main/java/com/example/ecommerce/repository/ItemEntityRepository.java
@@ -3,6 +3,8 @@ package com.example.ecommerce.repository;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
 public class ItemEntityRepository {
 
@@ -13,7 +15,7 @@ public class ItemEntityRepository {
         this.jdbc = jdbc;
     }
 
-    public void addRow(String name, String description, String imagePath, int price, boolean inStock) {
+    public void addRow(String name, String description, String imagePath, BigDecimal price, boolean inStock) {
         String itemSql = "INSERT INTO Item(name, description, imagePath, price, inStock) VALUES(?, ?, ?, ?, ?)";
         jdbc.update(itemSql, name, description, imagePath, price, inStock);
     }
