+++ src/main/java/com/example/ecommerce/repository/OrderEntityRepository.java
@@ -4,6 +4,8 @@ package com.example.ecommerce.repository;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
 public class OrderEntityRepository {
 
@@ -14,7 +16,7 @@ public class OrderEntityRepository {
         this.jdbc = jdbc;
     }
 
-    public void addRow(int userid, String orderDate, int subtotal, int shippingPrice, int taxes, int grandTotal, String shippingSelection) {
+    public void addRow(int userid, String orderDate, BigDecimal subtotal, BigDecimal shippingPrice, BigDecimal taxes, BigDecimal grandTotal, String shippingSelection) {
         String orderSql = "INSERT INTO 'Order'(userid, orderDate, subTotal, shippingPrice, taxes, GrandTotal, shippingSelection) VALUES (?, ?, ?, ?, ?, ?, ?, ?)";
         jdbc.update(orderSql, userid, orderDate, subtotal, shippingPrice, taxes, grandTotal);
     }
